+++ src/Collections/Environments.php
@@ -23,25 +23,33 @@ class Environments extends SiteOwnedCollection
     /**
      * Creates a multidev environment
      *
-     * @param string $to_env_id Name of new the environment
-     * @param Environment $from_env Environment to clone from
+     * @param string $target_env_id Name of new the environment
+     * @param Environment $source_env Environment to clone from
+     * @param array $options
+     *     bool no-db Do not copy database from the source environment
+     *     bool no-files Do not copy files from the source environment
      * @return Workflow
      */
-    public function create($to_env_id, Environment $from_env)
+    public function create($target_env_id, Environment $source_env, array $options = [])
     {
+        $params = [
+            'clone_database' => ['from_environment' => $source_env->id,],
+            'clone_files' => ['from_environment' => $source_env->id,],
+            'annotation' => "Create the \"{$target_env_id}\" environment.",
+        ];
+        if (isset($options['no-db']) && $options['no-db']) {
+            unset($params['clone_database']);
+        }
+        if (isset($options['no-files']) && $options['no-files']) {
+            unset($params['clone_files']);
+        }
+
         $workflow = $this->getSite()->getWorkflows()->create(
             'create_cloud_development_environment',
             [
                 'params' => [
-                    'environment_id' => $to_env_id,
-                    'deploy' => [
-                        'clone_database' => ['from_environment' => $from_env->id,],
-                        'clone_files' => ['from_environment' => $from_env->id,],
-                        'annotation' => sprintf(
-                            'Create the "%s" environment.',
-                            $to_env_id
-                        ),
-                    ],
+                    'environment_id' => $target_env_id,
+                    'deploy' => $params,
                 ],
             ]
         );